Updates to float, better icon and description

+++ src/Widgets/SpaceWidget/SpaceWidgetClass.js
@@ -2,7 +2,7 @@ import * as Scrivito from "scrivito";
 
 const SpaceWidget = Scrivito.provideWidgetClass("SpaceWidget", {
   attributes: {
-    size: "integer",
+    size: "float",
   },
 });
